+++ src/components/Error/style.js
@@ -1,16 +1,17 @@
-import styled from "styled-components";
+import styled from 'styled-components'
 
 export const ErrorWrapper = styled.div`
   position: fixed;
   z-index: 1;
   left: 0;
   top: 0;
+  display: flex;
+  flex-flow: column wrap;
+  justify-content: center;
   width: 100%;
-  height: 100%;
-  overflow: auto;
-  background-color: rgb(0, 0, 0);
+  height: 100vh;
   background-color: rgba(0, 0, 0, 0.4);
-`;
+`
 
 export const ErrorModal = styled.div`
   background-color: #fefefe;
@@ -19,28 +20,7 @@ export const ErrorModal = styled.div`
   border: 1px solid #888;
   width: 80%;
   max-width: 400px;
+  height: auto;
   color: black;
   text-align: center;
-`;
-
-export const RedirectButton = styled.a`
-  padding: 10px 20px;
-  background-color: #157bcf;
-  cursor: pointer;
-  text-decoration: none;
-  color: #fff;
-`;
-
-export const CloseBtn = styled.button`
-  color: #aaa;
-  float: right;
-  font-size: 28px;
-  font-weight: bold;
-`;
-
-export const ErrorTitle = styled.span`
-  color: #000;
-  font-size: 24px;
-  font-weight: bold;
-  display: block;
-`;
+`
