+++ src/containers/credit-line/Reducer.js
@@ -1,0 +1,10 @@
+import { combineReducers } from "redux";
+import panVerificationReducer from "./pan-details/Reducer";
+import personalDetails from "./PersonalDetails/Reducer";
+
+const rootReducer = combineReducers({
+  personalDetails,
+  panVerificationReducer
+});
+
+export default rootReducer;
